localiser-ui/requirements.txt
django>=2.0,<3.0
django-simple-history==2.7.3
